refactor: centralize owner server config reads
