+++ docs/architecture/migration-progress.md
@@ -5,18 +5,18 @@ Status values: `[ ]` not started, `[~]` in progress, `[x]` complete, `[!]` block
 ## Current Context
 
 - Issue: [`rustfs/backlog#660`](https://github.com/rustfs/backlog/issues/660)
-- Branch: `overtrue/arch-iam-credential-boundary`
-- Baseline: completed `C-011/C-012/C-013/API-055/API-059/API-079/API-080/API-081/API-082/API-083/API-084/API-085/API-086/API-087/API-088/API-089/API-090/API-091/API-092/API-093/API-094/API-095/API-096/API-097/API-098/API-099/API-100/API-101/API-102/API-103/API-104/API-105/API-106/API-107/API-108/API-109/API-110/API-111/API-112/API-113/API-114/API-115/API-116/API-117/API-118/API-119/API-120/API-121/API-122/API-123/API-124/API-125/API-126/API-127/API-128/API-129/API-130/API-131/API-132/API-133/API-134/API-135/API-136/API-137/API-138/API-139/API-140/API-141/API-142/API-143/API-144/API-145/API-146/API-147/API-148/API-149/API-150/API-151/API-152/API-153/API-154/API-155/API-156/API-157/API-158/API-159/API-160/API-161/API-162/API-163/API-164/API-165/API-166/API-167/API-168/API-169/API-170/API-171/API-172/API-173/API-174/API-175/API-176/API-177/API-178/API-179/API-180/API-181`.
-- Based on: API-180 prepared in PR #3791; this branch batches IAM root
-  credential boundary cleanup on top of that branch.
+- Branch: `overtrue/arch-owner-server-config-boundaries`
+- Baseline: completed `C-011/C-012/C-013/API-055/API-059/API-079/API-080/API-081/API-082/API-083/API-084/API-085/API-086/API-087/API-088/API-089/API-090/API-091/API-092/API-093/API-094/API-095/API-096/API-097/API-098/API-099/API-100/API-101/API-102/API-103/API-104/API-105/API-106/API-107/API-108/API-109/API-110/API-111/API-112/API-113/API-114/API-115/API-116/API-117/API-118/API-119/API-120/API-121/API-122/API-123/API-124/API-125/API-126/API-127/API-128/API-129/API-130/API-131/API-132/API-133/API-134/API-135/API-136/API-137/API-138/API-139/API-140/API-141/API-142/API-143/API-144/API-145/API-146/API-147/API-148/API-149/API-150/API-151/API-152/API-153/API-154/API-155/API-156/API-157/API-158/API-159/API-160/API-161/API-162/API-163/API-164/API-165/API-166/API-167/API-168/API-169/API-170/API-171/API-172/API-173/API-174/API-175/API-176/API-177/API-178/API-179/API-180/API-181/API-182`.
+- Based on: latest `origin/main` after API-181 content landed; this branch
+  batches IAM and scanner server-config boundary cleanup on top of main.
 - PR type for this branch: `consumer-migration`
 - Runtime behavior changes: none.
 - Rust code changes: route replication pool, outbound TLS generation, runtime
   region, KMS encryption service, runtime support handles, S3 Select DB,
   internode RPC metrics, IAM authorization/handler reads, notification
-  rules/event dispatch, admin OIDC/token-signing reads, and IAM root
-  credential consumers through AppContext-first or IAM-owned resolver
-  boundaries.
+  rules/event dispatch, admin OIDC/token-signing reads, IAM root credential
+  consumers, IAM OIDC config reads, and scanner runtime-config reads through
+  AppContext-first or owner-crate resolver boundaries.
 - CI/script changes: lock completed owner and test/fuzz boundaries against
   bare/glob imports, scattered raw ECStore facade subpaths, and startup
   runtime/root-server/table/S3/app shared/app bucket/app ECStore/admin facade
@@ -25,7 +25,7 @@ Status values: `[ ]` not started, `[~]` in progress, `[x]` complete, `[!]` block
   and storage owner thin bridge regressions, plus app context and notify
   event-bridge thin module regressions; accept the reviewed AppContext resolver
   reverse dependencies in the layer baseline.
-- Docs changes: record the API-136 through API-181 owner facade cleanup.
+- Docs changes: record the API-136 through API-182 owner facade cleanup.
 
 ## Phase 0 Tasks
 
@@ -4619,6 +4619,21 @@ Status values: `[ ]` not started, `[~]` in progress, `[x]` complete, `[!]` block
     migration guard, layer guard, diff hygiene, residual direct credential scan,
     Rust risk scan, branch freshness check, and three-expert review.
 
+- [x] `API-182` Centralize owner server config reads behind crate boundaries.
+  - Do: route IAM OIDC server-config reads through an IAM-owned helper and
+    scanner runtime-config global reads through scanner-owned resolver
+    functions.
+  - Acceptance: production IAM OIDC and scanner runtime-config consumers no
+    longer call the server-config global directly outside owner boundary
+    modules, while parser fallback and effective-provider behavior remain
+    unchanged.
+  - Must preserve: OIDC provider parsing, config fallback handling, scanner
+    runtime config defaults, warning logs, and runtime config refresh behavior.
+  - Verification: IAM/scanner compile coverage, IAM/scanner focused tests,
+    formatting, migration guard, layer guard, diff hygiene, residual server
+    config scan, Rust risk scan, branch freshness check, pre-commit quality
+    gate, and three-expert review.
+
 ## Next PRs
 
 1. `consumer-migration`: continue reducing direct global reads behind AppContext resolver boundaries.
@@ -4718,10 +4733,33 @@ Status values: `[ ]` not started, `[~]` in progress, `[x]` complete, `[!]` block
 | Quality/architecture | pass | API-181 keeps IAM root credential reads centralized in an IAM-owned helper and exposes only a root access-key predicate to protocol code. |
 | Migration preservation | pass | Root owner checks, legacy IAM decrypt fallback, token signing, and gateway policy args preserve existing credential semantics. |
 | Testing/verification | pass | IAM/protocol compile, IAM unit tests, formatting, migration/layer guards, diff hygiene, residual credential scan, and Rust risk scan passed for API-181. |
+| Quality/architecture | pass | API-182 keeps server-config global reads inside IAM and scanner owner boundary helpers without widening public APIs. |
+| Migration preservation | pass | OIDC provider parsing and scanner runtime-config fallback semantics keep the same inputs, defaults, and warning behavior. |
+| Testing/verification | pass | IAM/scanner compile, focused tests, formatting, migration/layer guards, diff hygiene, residual server-config scan, Rust risk scan, and pre-commit passed for API-182. |
 
 ## Verification Notes
 
 Passed before push:
+
+- Issue #660 API-182 current slice:
+  - `cargo check -p rustfs-iam -p rustfs-scanner --tests`: passed.
+  - `cargo test -p rustfs-iam --lib`: passed.
+  - `cargo test -p rustfs-scanner --lib -- --test-threads=1`: passed.
+  - `cargo fmt --all`: passed.
+  - `cargo fmt --all --check`: passed.
+  - `git diff --check`: passed.
+  - `bash -n scripts/check_architecture_migration_rules.sh`: passed.
+  - `./scripts/check_architecture_migration_rules.sh`: passed.
+  - `./scripts/check_layer_dependencies.sh`: passed.
+  - IAM/scanner server-config scan: passed; touched production direct
+    `get_global_server_config` calls are isolated to IAM and scanner boundary
+    modules, while existing app context, config, and ECStore owner reads remain
+    unchanged.
+  - Rust risk scan: no new production unwrap/expect, panic/todo/unsafe, or cast
+    risk.
+  - Branch freshness check: rebased on latest `origin/main` after PR #3789 and
+    API-181 content landed upstream.
+  - `make pre-commit`: passed.
 
 - Issue #660 API-181 current slice:
   - `cargo check -p rustfs-iam -p rustfs-protocols --tests`: passed.
